+++ src/config/images.ts
@@ -1,0 +1,10 @@
+const defaultPlugins: {
+  gif?: boolean;
+  jpeg?: boolean;
+  png?: boolean;
+  svg?: boolean;
+} = {};
+
+export default {
+  defaultPlugins
+};